CASMPET-3814: Configure istio proxy sidecar with ImagePullPolicy IfNotPresent
Currently the sidecar injector is setting the ImagePullPolicy to
Always. This makes the proxyv2 image caching not work as expected
since just about every pod is going to try to access nexus to get
this image even if it's cached, and will fail if nexus is down.

This just configures istio such that the sidecar injector will
use IfNotPresent ImagePullPolicy so the request will be fully
handled by the cached image.