CASMPET-3836: Increase timeouts for OPA, remove "/apis/rm", FQDN hostname
I was seeing lots of 503 UAEX failures on a system where the time
for the request was ~5s. We'd seen this before but I'd increased
the timeout in the envoyfilter to 10s. That change happened before
I had to refactor JWT cert handling to move it from istio into
OPA.

Google turned up that the http.send function has a default 5s
timeout[0]. I tried increasing the timeout on the system where the
failures were happening and after this the 503 errors went away.

[0] https://github.com/open-policy-agent/opa/issues/2099

This change adds a variable to set the http timeout and defaults
it to 10s.

I also increased the default envoyfilter timeout to 25s since
there are potentially 2 calls to http.send in the OPA rules.

In addition I removed the /apis/rm paths from the auth checks
because this service was actually removed (CASMSEC-307).

Also, added the . to the DNS names in the JWKS query URLs. This
prevents the resolver from going through the DNS search path so
there's fewer DNS queries.

CASMPET-3838 : SCALE: CSM 0.8.11: cray-opa pods being OOM killed at 2GiB

(cherry picked from commit 8d71d3da16e3e6c45932a549cd91418031fef0c7)