CASMPET-3835: Adjust spire-jwks/keys response so OPA caches response
OPA wasn't caching the spire-jwks/keys response and was instead
requesting the keys again on every request. The intent was to
cache this response in order to prevent making a new request for
every request that came in.

This change adjusts the VirtualService for spire-jwks so that it
sets some caching headers on the certs response the same as the
Keycloak certs response which is getting cached (and also uses
a VS to adjust the response headers).

Also includes:

CASMSEC-313 Bump postgres pvc to 20G

spire-server causes a lot of logs during the compute boot process. Upping the postgres database pvc to 20GB should allow enough space to hold the logs.